moved formatting stuff into its own class
